// src/api/generated/models/metricResponseDto.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * Requirements Backend API
 * HTTP API for the Requirements Management app.
 * OpenAPI spec version: 0.0.1
 */

export interface MetricResponseDto {
  id: string;
  projectId: string;
  /** @pattern ^MET-[0-9]{4}$ */
  key: string;
  value: string;
  /** @nullable */
  description: string | null;
  createdAt: string;
  updatedAt: string;
}